tech/pyproject.toml
[tool.poetry]
name = "tech"
version = "0.1.0"
description = ""
authors = ["KauanAdvolveAi <kauan.silva@advolve.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
uvloop = "^0.17.0"
httptools = "^0.5.0"
fastapi = {extras = ["standard"], version = "^0.114.0"}
sqlalchemy = "^2.0.34"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
psycopg = {extras = ["binary"], version = "^3.2.2"}
psycopg-binary = "^3.2.4"
boto3 = "^1.37.4"
jose = "^1.0.0"
jwt = "^1.3.1"
pyjwt = "^2.10.1"
behave = "^1.2.6"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
testcontainers = "^4.8.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev tech/adapters/driver/http/app.py'


test = 'pytest --cov=tech -vv'
post_test = 'coverage html'

lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . ; ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
